# senotype-api-spec.yaml
openapi: 3.0.3
info:
  description: "The SenNet Senotype API is a RESTful web service with create, read, update, and delete operations for the SenNet senotype data."
  title: SenNet Senotype API
  version: 1.0.0
  contact:
    name: SenNet Help Desk
    email: help@sennetconsortium.org
  license:
    name: MIT License
    url: "https://github.com/sennetconsortium/senotype-api/blob/main/license.txt"
servers:
  - url: "https://senotype.api.sennetconsortium.org"

tags:
  - name: Status
    description: Health check endpoints
  - name: Senotypes
    description: CRUD operations for senotypes
  - name: Valuesets
    description: Valueset lookup endpoints

paths:
  /status:
    get:
      tags:
        - Status
      summary: Health check
      operationId: getStatus
      responses:
        "200":
          description: Service is running
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

  /senotypes:
    get:
      tags:
        - Senotypes
      summary: Get all senotypes
      operationId: getSenotypes
      description: >
        Requires the authenticated user to be a member of at least one of the
        `senotype-curate`, `senotype-edit`, or `senotype-publish` Globus groups.
      security:
        - bearerAuth: []
      parameters:
        - name: return_dict
          in: query
          required: false
          schema:
            type: string
            enum: ["true", "false"]
            default: "true"
          description: Whether to return the full senotype objects
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
          description: Maximum number of senotypes to return
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Number of senotypes to skip before collecting the result set
        - name: order
          in: query
          required: false
          schema:
            type: string
            enum: ["asc", "desc"]
            default: "asc"
          description: >
            Sort direction applied to senotypes by creation time (with UUID
            as a tiebreaker) before pagination is applied
      responses:
        "200":
          description: List of senotypes
          content:
            application/json:
              schema:
                type: object
                properties:
                  senotypes:
                    type: array
                    items:
                      type: object
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      tags:
        - Senotypes
      summary: Create a senotype
      operationId: createSenotype
      description: >
        Requires the authenticated user to be a member of at least one of the
        `senotype-curate`, `senotype-edit`, or `senotype-publish` Globus groups.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SenotypeRequest"
      responses:
        "201":
          description: Senotype created successfully
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /senotypes/{uuid}:
    get:
      tags:
        - Senotypes
      summary: Get a senotype by UUID
      operationId: getSenotype
      description: >
        Requires the authenticated user to be a member of at least one of the
        `senotype-curate`, `senotype-edit`, or `senotype-publish` Globus groups.
      security:
        - bearerAuth: []
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
            pattern: "^[a-f0-9]{32}$"
            example: 73df6847495fcd61e4b6e5292ab2c9dc
      responses:
        "200":
          description: Senotype found
          content:
            application/json:
              schema:
                type: object
                properties:
                  senotype:
                    type: object
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    put:
      tags:
        - Senotypes
      summary: Update a senotype
      operationId: updateSenotype
      description: >
        Requires the authenticated user to be a member of the
        `senotype-curate` or `senotype-edit` Globus group. Members of
        `senotype-curate` may update any senotype; members of
        `senotype-edit` may only update senotypes they created.
      security:
        - bearerAuth: []
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
            pattern: "^[a-f0-9]{32}$"
            example: 73df6847495fcd61e4b6e5292ab2c9dc
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SenotypeRequest"
      responses:
        "200":
          description: Senotype updated successfully
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      tags:
        - Senotypes
      summary: Delete a senotype
      operationId: deleteSenotype
      description: >
        Requires the authenticated user to be a member of the
        `senotype-curate` or `senotype-edit` Globus group. Members of
        `senotype-curate` may delete any senotype; members of
        `senotype-edit` may only delete senotypes they created.
      security:
        - bearerAuth: []
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
            pattern: "^[a-f0-9]{32}$"
            example: 73df6847495fcd61e4b6e5292ab2c9dc
      responses:
        "200":
          description: Senotype deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Senotype deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /valuesets:
    get:
      tags:
        - Valuesets
      summary: Get valuesets
      operationId: getValuesets
      parameters:
        - name: predicate_term
          in: query
          required: false
          schema:
            type: string
            enum:
              - assay
              - context
              - hallmark
              - inducer
              - located_in
              - microenvironment
              - sex
              - taxon
          description: Filter valuesets by predicate term
      responses:
        "200":
          description: List of valuesets
          content:
            application/json:
              schema:
                type: object
                properties:
                  valuesets:
                    type: array
                    items:
                      $ref: "#/components/schemas/Valueset"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        Globus groups bearer token. Different operations require membership
        in one or more of the `senotype-curate`, `senotype-edit`, or
        `senotype-publish` Globus groups; see each operation's description
        for the specific requirements.

  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            type: object
            properties:
              errors:
                type: object
                additionalProperties:
                  type: array
                  items:
                    type: string
            example:
              errors:
                taxon: ["Valueset 'NCBI:9999' not found in valuesets"]
                organ: ["Organ 'UBERON:0000000' not found in UBKG"]
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "limit must be between 1 and 100"
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string

    Pagination:
      type: object
      properties:
        total:
          type: integer
          description: Total number of senotypes matching the query
          example: 137
        limit:
          type: integer
          description: Maximum number of senotypes returned in this response
          example: 25
        offset:
          type: integer
          description: Number of senotypes skipped before this response's result set
          example: 0
        order:
          type: string
          enum: ["asc", "desc"]
          description: Sort direction applied before pagination
          example: "asc"

    BMI:
      type: object
      required:
        - value
        - unit
      properties:
        value:
          type: number
          example: 22.5
        unit:
          type: string
          enum: [kg/m^2]
        lowerbound:
          type: number
          nullable: true
          minimum: 0
          exclusiveMinimum: true
          example: 18.5
        upperbound:
          type: number
          nullable: true
          minimum: 0
          exclusiveMinimum: true
          example: 30.0

    Age:
      type: object
      required:
        - value
        - unit
      properties:
        value:
          type: number
          example: 65
        unit:
          type: string
          enum: [year]
        lowerbound:
          type: number
          nullable: true
          minimum: 0
          exclusiveMinimum: true
          example: 60
        upperbound:
          type: number
          nullable: true
          minimum: 0
          exclusiveMinimum: true
          example: 70

    Diagnosis:
      type: object
      required:
        - code
        - term
      properties:
        code:
          type: string
          example: DOID:9351
        term:
          type: string
          example: diabetes mellitus

    RegulatedMarker:
      type: object
      required:
        - action
        - marker
      properties:
        action:
          type: string
          enum:
            - up_regulates
            - down_regulates
            - inconclusively_regulates
        marker:
          oneOf:
            - type: string
              pattern: '^HGNC:\d+$'
              description: HGNC gene identifier (human)
            - type: string
              pattern: '^MGI:\d+$'
              description: MGI gene identifier (mouse)
            - type: string
              pattern: "^UNIPROTKB:[A-Z0-9]+$"
              description: UniProtKB protein identifier
          example: HGNC:11027

    SenotypeRequest:
      type: object
      required:
        - title
        - description
        - taxon
        - organ
        - cell_type
        - hallmark
      properties:
        title:
          type: string
          minLength: 1
          example: Example Senotype
        description:
          type: string
          minLength: 1
          example: This is an example senotype for testing purposes.
        taxon:
          type: array
          minItems: 1
          items:
            type: string
          example: ["NCBI:9606"]
        organ:
          type: array
          minItems: 1
          items:
            type: string
          example: ["UBERON:0001066"]
        cell_type:
          type: array
          minItems: 1
          items:
            type: string
            pattern: '^CL:\d+$'
          example: ["CL:0000507"]
        hallmark:
          type: array
          minItems: 1
          items:
            type: string
          example: ["SENOTYPE_VS:C00209"]
        bmi:
          $ref: "#/components/schemas/BMI"
          nullable: true
        age:
          $ref: "#/components/schemas/Age"
          nullable: true
        microenvironment:
          type: array
          nullable: true
          items:
            type: string
          example: ["SENOTYPE_VS:C00302"]
        inducer:
          type: array
          nullable: true
          items:
            type: string
          example: ["SENOTYPE_VS:C00405"]
        assay:
          type: array
          nullable: true
          items:
            type: string
          example: ["SENOTYPE_VS:C00507"]
        sex:
          type: array
          nullable: true
          items:
            type: string
          example: ["SNOMEDCT_US:248153007"]
        diagnosis:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/Diagnosis"
        citation:
          type: array
          nullable: true
          items:
            type: string
            pattern: '^PMID:\d+$'
          example: ["PMID:41705430"]
        origin:
          type: array
          nullable: true
          items:
            type: string
            pattern: "^RRID:"
          example: ["RRID:scr_020638"]
        dataset:
          type: array
          nullable: true
          items:
            oneOf:
              - type: string
                pattern: "^[a-f0-9]{32}$"
                description: Dataset UUID
              - type: string
                pattern: '^SNT\d{3}\.[A-Z]{4}\.\d{3}$'
                description: Dataset SenNet ID
          example: ["73df6847495fcd61e4b6e5292ab2c9dc", "SNT123.ABCD.456"]
        specified_marker_set:
          type: array
          nullable: true
          items:
            oneOf:
              - type: string
                pattern: '^HGNC:\d+$'
                description: HGNC gene identifier (human)
              - type: string
                pattern: '^MGI:\d+$'
                description: MGI gene identifier (mouse)
              - type: string
                pattern: "^UNIPROTKB:[A-Z0-9]+$"
                description: UniProtKB protein identifier
          example: ["HGNC:9561"]
        regulated_marker_set:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/RegulatedMarker"
      example:
        title: Example Senotype
        description: This is an example senotype for testing purposes.
        bmi:
          value: 22.5
          unit: kg/m^2
          lowerbound: 18.5
          upperbound: 30.0
        age:
          value: 65
          unit: year
          lowerbound: 60
          upperbound: 70
        taxon: ["NCBI:9606"]
        organ: ["UBERON:0001066"]
        cell_type: ["CL:0000507"]
        microenvironment: ["SENOTYPE_VS:C00302"]
        inducer: ["SENOTYPE_VS:C00405"]
        hallmark: ["SENOTYPE_VS:C00209"]
        assay: ["SENOTYPE_VS:C00507"]
        sex: ["SNOMEDCT_US:248153007"]
        diagnosis:
          - code: DOID:9351
            term: diabetes mellitus
        citation: ["PMID:41705430"]
        origin: ["RRID:scr_020638"]
        dataset: ["73df6847495fcd61e4b6e5292ab2c9dc", "SNT123.ABCD.456"]
        specified_marker_set: ["HGNC:9561"]
        regulated_marker_set:
          - action: down_regulates
            marker: HGNC:11027
          - action: up_regulates
            marker: UNIPROTKB:Q13201

    Valueset:
      type: object
      required:
        - code
        - term
        - predicate_term
      properties:
        code:
          type: string
          example: NCBI:9606
        term:
          type: string
          example: human
        predicate_term:
          type: string
          example: in_taxon
        predicate_iri:
          type: string
          nullable: true
          example: http://purl.obolibrary.org/obo/RO_0002162